pool delete route complete

+++ api/routers/pools.py
@@ -66,9 +66,21 @@ def get_pools_details(
 @router.delete("/api/pools/{pool_id}")
 def delete_pool(
     pool_id: int,
+    poolowner_id: int,
     user: UserResponse = Depends(try_get_jwt_user_data),
-    queries: PoolQueries = Depends()
+    queries: PoolQueries = Depends(),
+    pool_amenities_queries: PoolAmenitiesQueries = Depends()
 ):
+    # look for pool and throw 404 (get by pool owner id)
+    # delete all associations line 75, get all ids and delete in for loop (delete)
+    # for pool id giving amenity id as well
+    if not poolowner_id:
+        raise HTTPException(
+            status_code=404, detail="You are not the pool owner"
+        )
+    amenity_list = pool_amenities_queries.get_pool_with_amenities(pool_id=pool_id)
+    for amenity_ids in amenity_list:
+        pool_amenities_queries.delete(amenity_id=amenity_ids, pool_id=pool_id)
     return {
         "success": queries.delete_pool(pool_id=pool_id, poolowner_id=user.id)
     }
